Remove `artifacts.yaml` part in favor of `artifacts:` section in DVC (#346)
Related to https://github.com/iterative/dvc/issues/9219

One question I'd like to discuss: instead of removing `gto describe`
(and maybe `gto annotate` and `gto remove`), we could keep them. At
least for `describe` it's pretty trivial to support - we can call DVC
API to read annotations.

Also this can be helpful if we're planning to implement the command in
DVC - here we can implement and check it works and satisfies user needs.