fix(registry): push to remote automatically only on cloned repos (#417)
Fixes #405 

We've changed the semantic of the operation during the recent migration
to `scmrepo`. The key thing is the `has_remote(reg.scm)` calls and
implementation. I think the original intention was to push automatically
when we run GTO operations on a remote repo (means we are cloning it
into a temp dir, do some op, and push the result back).

## TODO

- [x] Tests
- [x] Review CLI option description
- [x] Review docs

## Docs

Relevant docs update is here
https://github.com/iterative/dvc.org/pull/4879